Reestructure some other files
